requirements.txt
# !!!! Self generated file Do not edit this !!!!
uvicorn==0.18.3
numpy==1.23.4
matplotlib==3.6.2
seaborn==0.12.1
notebook==6.5.2
pandas==1.5.1
scikit-learn==1.1.3
SQLAlchemy==1.4.23
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
requests==2.29.0
